Revert "dri2: Don't make reference to noClientException"
It's true that the value would always be -1, if it's not zero, but it's
usually zero is the problem. As a result we return failure from
otherwise successful indirect GLX paths, which isn't very nice of us.

This reverts commit 7d33ab0f8c7958b205076f71e4b47c24aace77fd.

Fixes: https://gitlab.freedesktop.org/xorg/xserver/issues/211
(cherry picked from commit e1fa3beb2fe2519e69f859f0acdc68e5a770de27)